+++ src/inotify.rs
@@ -8,14 +8,7 @@ use std::{
 use nix::sys::inotify::Inotify;
 use tokio::io::{unix::AsyncFd, Interest};
 
-use crate::new::{
-    external::error::{AnotifyError, AnotifyErrorType, Result},
-    EventFilter,
-};
-
 use super::binding::{Binding, BindingEvent, BindingEventType};
-
-mod stats;
 
 struct OwnedInotify(Inotify);
 
@@ -47,7 +40,7 @@ impl AsRawFd for OwnedInotify {
 /// Platform bindings for [`Inotify`][`nix::sys::inotify::Inotify`]
 pub struct InotifyBinding {
     fd: AsyncFd<OwnedInotify>,
-    stats: stats::Stats,
+    stats: Stats,
 }
 
 #[repr(transparent)]
@@ -275,5 +268,70 @@ impl Binding for InotifyBinding {
                 events.push(Self::convert_event(event));
             }
         }
+    }
+}
+
+struct Stats {
+    span: tracing_impl::Span,
+}
+
+impl Stats {
+    pub fn new() -> Self {
+        let span = tracing_impl::trace_span!(
+            "runtime.resource",
+            concrete_type = "Inotify",
+            kind = "file",
+            is_internal = false,
+            inherits_child_attrs = true,
+        );
+
+        span.in_scope(|| {
+            tracing_impl::trace!(
+                target: "runtime::resource::state_update",
+                watching = 0,
+                watching.unit = "files",
+                watching.op = "override"
+            );
+
+            tracing_impl::trace!(
+                target: "runtime::resource::state_update",
+                events = 0,
+                events.unit = "events",
+                events.op = "override"
+            );
+        });
+
+        Self { span }
+    }
+
+    pub fn inc_files(&self, count: usize) {
+        self.span.in_scope(|| {
+            tracing_impl::trace!(
+                 target: "runtime::resource::state_update",
+                 watching = count,
+                 watching.unit = "files",
+                 watching.op = "add"
+            );
+        });
+    }
+
+    pub fn dec_files(&self, count: usize) {
+        self.span.in_scope(|| {
+            tracing_impl::trace!(
+                 target: "runtime::resource::state_update",
+                 watching = count,
+                 watching.unit = "files",
+                 watching.op = "sub"
+            );
+        });
+    }
+
+    pub fn note_events(&self, count: usize) {
+        tracing_impl::trace!(
+            target: "runtime::resource::state_update",
+            events = count,
+            events.unit = "events",
+            events.op = "add"
+        );
     }
 }
