--- src/new.rs
@@ -1,4 +1,0 @@
-mod external;
-mod internal;
-
-pub use external::*;
