+++ src/registry.rs
